Move "levels" parameter to the ll_rayshooter_t struct
This makes a new function ll_rayshoot necessary to start the iteration
with the level given as an additional parameter, enabling us to fix a
off-by-one issue in the levels parameter (before you actually had to
give the number of levels minus one as a parameter).

Additionally, ll_rayshoot_bilinear is not exported anymore to the
library.  It does not seem very useful as a standalone function, and
if you really want, you can still call it indirectly via ll_rayshoot
(setting levels to 0).