Revert "New heuristics do not need "overshoot" anymore"
This reverts commit b4ef82d85fbda579c7c75596d3bea012c69bee2c.  There
are corner cases where the overshoot seems to be necessary.  I fiddled
around with relaxing the heuristics to reduce the number of subpatches
that should be reshot, but I could not significantly improve
performance.  Thus I stick to the heuristics that shoots the widest
region because it is the least prone to artifacts.

Conflicts:

	ll.c