First operational version of triangulated ray shooting
Ray shooting using anti-aliased rendering of triangles on the finest
level (so it is actually rather "triangle shooting" than "ray
shooting").  This version is mainly copied from some old anti-aliasing
test code and not yet optimised for the case at hand.  It runs at
about half the speed of the bilinearly interpolated ray shooting, but
is less prone to moiré.