Add a simple example for the use of the Python interface
The script animation.py computes a sequence of magnification patterns,
showing the transition from the close separation via the intermediate
to the wide separation case.  The output is RAW greyscale data.  The
script rawtogif.sh can be used to generate an animated GIF from the
files (which can be viewed in most web browsers).