Move levels property from rayhooter to pacthes struct
It has always been a peculiarity that the levels property was a field
of the rayshooter struct.  It actually corresponds to properties like
the shooting rectangle, xrays and yrays.  Now it is treated in the
same way as the latter.  This also flattens out some strangeness in
the Python code.

This patch strangely boosts the performance of multi-threaded ray
shooting from the Python interface by up to a factor of two.  No idea
how this comes about.