Fixing FITSReader: endianness, liberality and lens table extension name
The FITSReader processor now converts the lens list to the machine's
native endianness to make it suitable to be passed to C code.
Furthermore, FITSReader now is more liberal regarding header fields or
the lens table:  if they are not in the file, the corresponding fields
simply are not exported.  The lens table now gets the extension name
"lenses" for easy identification.