Make export of the original source profile optional
Introduce the switch "export_kernel" to decide whether the original source
profile should be included in the results of SourceProfile.run().

GllSourceProfile got a set_config() method as a hack to make old .gll files
work.  A real solution how to cope with format changes is needed.
Probably we should drop the current pickle format and use something less
insane.