Bulk rename: magpattern -> magpat, drop suffix "_t" on struct names
There has been a mix between the abbreviations "magpat" and "magpattern"
for magnification pattern, which was also visible in parts of the
interface.  This is now cleaned up in favour of the shorter "magpat".

Furthermore, the struct names in the C code have lost their "_t" suffix.
The names are in the struct namespace anyway, they don't need a suffix to
mark them as types.  The suffix would only make sense for typedefed names.