Make BasicRayshooter.finalise_subpathces() work with new Magpat class
This has been a very strange bug: if "a" and "b" are plain NumPy arrays,

    a += b

will add b to a, and a stays the same object.  If they are Magpat instances
though, "a" will become a new view to the same data.  Here, this caused
that the dtype of the first magpat was not reset after adding the patterns.